requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap4==2022.1
cron-descriptor==1.4.0
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2.4
django-allauth==0.54.0
django-celery-beat==2.5.0
django-crispy-forms==2.0
django-huey==1.1.1
django-timezone-field==5.1
idna==3.4
kombu==5.3.1
oauthlib==3.2.2
packaging==23.1
prompt-toolkit==3.0.39
pycparser==2.21
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
